Update UI and site rules
