+++ public/styles/sharedHtmlAndComponentStyles.css
@@ -1,7 +1,6 @@
 /* Styles that are shared with components */
-a, button {
+a {
   border-color: var(--clr-darker-blue);
-  border-radius: 5px;
   border-width: 1px;
   color: white;
   font-size: 1.5rem;
@@ -9,7 +8,7 @@ a, button {
   text-decoration: none;
 }
 
-label {
+b,h1,h2,label,li {
   color: var(--clr-darker-blue)
 }
 
@@ -22,7 +21,7 @@ label {
 }
 
 p,span {
-  color: var(--clr-dark-blue);
+  color: var(--clr-darker-blue);
   font-size: 1.5rem;
 }
