+++ src/beta/delete.html
@@ -13,8 +13,8 @@
   <title>Delete group?</title>
 
   <script type="module">
-    import {DeleteGroupPageComponent} from "../ui/groups/DeleteGroupPageComponent.ts";
-    customElements.define('delete-group-page-component',DeleteGroupPageComponent)
+    import {DeleteGroupComponent} from "../ui/groups/DeleteGroupComponent.ts";
+    customElements.define('delete-group-page-component',DeleteGroupComponent)
   </script>
 
 </head>
