+++ src/beta/login.html
@@ -17,7 +17,7 @@
     fetchpriority="low"
     type="module"
   >
-    import {LoginStatusComponent} from "../shared/components/LoginStatusComponent.ts";
+    import {LoginStatusComponent} from "../shared/html/LoginStatusComponent.ts";
     import {LoginComponent} from "../ui/auth/LoginComponent.ts";
 
     customElements.define("login-status-component", LoginStatusComponent);
