+++ src/conventions.html
@@ -10,7 +10,7 @@
   <link rel="preconnect" href="api.gatherspiel.com" />
 
   <script type="module">
-    import {ConventionListComponent} from "./ui/listing/components/ConventionListComponent.js";
+    import {ConventionListComponent} from "./ui/list/components/ConventionListComponent.js";
     customElements.define("convention-list-component",ConventionListComponent)
   </script>
