+++ src/gameRestaurants.html
@@ -12,7 +12,7 @@
   <script type="module">
     import {
       GameRestaurantListComponent,
-    } from "./ui/listing/components/GameRestaurantListComponent.js";
+    } from "./ui/list/components/GameRestaurantListComponent.js";
     customElements.define("game-restaurant-list-component",GameRestaurantListComponent)
   </script>
