+++ src/gameStores.html
@@ -10,7 +10,7 @@
   <link rel="preconnect" href="api.gatherspiel.com" />
 
   <script type="module">
-    import {GameStoreListComponent} from "./ui/listing/components/GameStoreListComponent.js";
+    import {GameStoreListComponent} from "./ui/list/components/GameStoreListComponent.js";
     customElements.define("game-store-list-component",GameStoreListComponent)
   </script>
