+++ src/shared/data/EventDataUtils.ts
