+++ src/shared/html/SiteRules.ts
@@ -1,0 +1,26 @@
+
+
+export function getSiteRulesHtml(){
+  return `    
+    <h1>Rules for groups and group events</h1>
+    <b>Note: These rules only apply to content posted on dmvboardgames.com</b>
+
+    <h2>General guidelines</h2>
+    <ul>
+      <li>All events must be in person at a location within the DMV area and must be related to board games.</li>
+      <li>Any links posted in descriptions must be informational content relevant to an event such as a group website
+          or details about board games. Content of the link must also be visible without logging in or entering personal information.</li>
+      <li>If organizers need to collect money for expenses related to hosting an event, they should share payment details or collect
+          money in person at the event.</li>
+      <li>Don't post information related to playtest games. If you are looking to playtest a game, it is recommended to reach
+        out to <a href="https://www.breakmygame.com/">Break My Game</a> or try to organize a playtest in person at an event.</li>
+    </ul>
+    <h2>Prohibited content</h2>
+      <ul>
+        <li>Contact information including email addresses, group chat information, or phone numbers.</li>
+        <li>Information related to exchanging money not related to necessary costs for the event. This also includes posting about buying or selling games,
+        or sharing payment service information.</li>
+        <li>Links to websites that involve spending money such as an online store </li>
+      </ul>
+    </ul>`
+}
